requirements.txt
easyocr==1.4.1
mouse==0.7.1
numpy==1.19.3
opencv_python_headless==4.5.5.62
Pillow==9.0.1
PySimpleGUI==4.57.0
pywin32==301
requests==2.26.0
win32gui==221.6
easyocr
yaml
requests